requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bson==0.5.10
dnspython==2.6.1
fastapi==0.112.2
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
